% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{download_spt_files}
\alias{download_spt_files}
\title{Download the latest files from a SharePoint folder}
\usage{
download_spt_files(recent_files, folder_name, directory, site_url)
}
\arguments{
\item{recent_files}{A data frame containing the file names to be downloaded. Typically, this would be
generated using a function like \code{list_recent_spt_files()}.}

\item{folder_name}{A character string specifying the SharePoint folder name from which to download the files.}

\item{directory}{A character string specifying the local directory where the files will be saved.}

\item{site_url}{A character string containing the SharePoint site URL.}
}
\value{
Downloads the specified files to the local directory. Returns nothing (NULL).
}
\description{
This function connects to a specified SharePoint folder and downloads the most recent files
to a specified local directory. The files are downloaded based on the file names provided in the
\code{recent_files} data frame, which should contain the names of the files to be downloaded.
}
\examples{
\dontrun{
  # List and download the most recent files from the specified SharePoint folder
  recent_files <- list_recent_spt_files(
    X = 14,
    folder_name = "Daily UEC Performance/Raw data/Recent data/ADA",
    site_url = "https://csucloudservices.sharepoint.com/sites/BSOLEmbeddedBI"
  )
  download_spt_files(
    recent_files = recent_files,
    folder_name = "Daily UEC Performance/Raw data/Recent data/ADA",
    directory = paste0(directory, "/Data/New ADA"),
    site_url = "https://csucloudservices.sharepoint.com/sites/BSOLEmbeddedBI"
  )
}
}
