% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_latest_excel_dataframes}
\alias{get_latest_excel_dataframes}
\title{Read the most recent Excel files into data frames}
\usage{
get_latest_excel_dataframes(latest_files)
}
\arguments{
\item{latest_files}{A character vector containing the full paths to the Excel files to be read.}
}
\value{
A list of data frames, where each data frame corresponds to an Excel file from \code{latest_files}.
}
\description{
This function takes a list of file paths to Excel files and reads each file into a data frame.
The resulting data frames are returned in a list. The first row of each Excel file is not used as
column names, and all columns are read without headers.
}
\details{
The \code{latest_files} argument, which is a list of file paths, can be obtained using the
\code{get_latest_excel_filenames()} function to first retrieve the most recent Excel files from a directory.
}
\examples{
\dontrun{
  # Get the latest filenames from the directory
  latest_filenames <- get_latest_excel_filenames(
    directory = "path/to/excel/files",
    X = 14
  )
  # Read the Excel files into data frames
  latest_dataframes <- get_latest_excel_dataframes(latest_filenames)
  # Access the first data frame
  print(latest_dataframes[[1]])
}
}
