% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{move_all_files}
\alias{move_all_files}
\title{Move all files from one folder to another within a main directory}
\usage{
move_all_files(main_directory, source_folder, destination_folder)
}
\arguments{
\item{main_directory}{A character string specifying the main directory that contains both the source
and destination folders.}

\item{source_folder}{A character string specifying the name of the source folder from which files will be moved.}

\item{destination_folder}{A character string specifying the name of the destination folder to which files will be moved.}
}
\value{
Returns \code{NULL} after moving all files from the source to the destination folder.
Prints a success message for each moved file.
}
\description{
This function moves all files from a specified source folder to a destination folder within the
main directory. If the destination folder does not exist, it is created. The function overwrites
any files in the destination folder if they already exist.
}
\examples{
\dontrun{
  # Move all files from 'New ADA' folder to 'Old ADA' folder within the main directory
  move_all_files(
    main_directory = file.path(directory, "Data"),
    source_folder = "New Data",
    destination_folder = "Old Data"
  )
}
}
