% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{process_ADA_data}
\alias{process_ADA_data}
\title{Process and combine ADA data frames}
\usage{
process_ADA_data(
  latest_dataframes,
  latest_filenames,
  latest_date,
  drop_na = TRUE
)
}
\arguments{
\item{latest_dataframes}{A list of data frames containing ADA data to be processed. Can be obtained using \code{get_latest_excel_dataframes()}.}

\item{latest_filenames}{A character vector of file names corresponding to the ADA data frames. Can be obtained using \code{get_latest_excel_filenames()}.}

\item{latest_date}{A Date object representing the cutoff date to filter the combined data.}

\item{drop_na}{A logical value indicating whether to drop rows with missing values in each data frame. Default is TRUE.}
}
\value{
A combined data frame containing processed ADA data, filtered by the specified date.
}
\description{
This function processes a list of ADA data frames and combines the processed results into a single data frame.
It extracts the date from the file name, processes the data (dropping NA values and converting necessary columns
to numeric), and combines the results into a single table, filtering by a provided date.
}
\details{
The \code{latest_dataframes} argument, which is a list of data frames, can be obtained using the
\code{get_latest_excel_dataframes()} function. The \code{latest_filenames} argument, which is a vector of file names,
can be obtained using the \code{get_latest_excel_filenames()} function.
}
\examples{
\dontrun{
  # Obtain the latest filenames and data frames
  latest_filenames <- get_latest_excel_filenames(directory = "path/to/excel/files", X = 14)
  latest_dataframes <- get_latest_excel_dataframes(latest_filenames)

  # Process and combine the ADA data
  combined_df <- process_ADA_data(
    latest_dataframes = latest_dataframes,
    latest_filenames = latest_filenames,
    latest_date = as.Date("2024-09-01")
  )
  print(combined_df)
}
}
